Split compile response storage responsibilities
