+++ Packages/src/Editor/FirstPartyTools/Compile/CompileResultRecordingContext.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9eb286c8b93d4608ba9cd34ad1177067
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
