--- Packages/src/Editor/FirstPartyTools/Compile/CompileSessionResultService.cs
@@ -1,206 +1,0 @@
-using System;
-using System.Diagnostics;
-using System.Linq;
-using Newtonsoft.Json;
-
-using io.github.hatayama.UnityCliLoop.Domain;
-using io.github.hatayama.UnityCliLoop.Infrastructure;
-using io.github.hatayama.UnityCliLoop.ToolContracts;
-
-namespace io.github.hatayama.UnityCliLoop.FirstPartyTools
-{
-    /// <summary>
-    /// Stores compile results in Editor SessionState so the CLI can read them after Domain Reload.
-    /// </summary>
-    internal static class CompileSessionResultService
-    {
-        private const string MissingTestFrameworkReferenceHint =
-            "Possible test asmdef issue: Unity test framework symbols are missing. Make sure com.unity.test-framework is installed and add optionalUnityReferences: [\"TestAssemblies\"] or enable testAssemblies on the test asmdef.";
-
-        internal static CompileResponse CreateCompileResult(
-            CompileResult result,
-            bool forceRecompile)
-        {
-            Debug.Assert(result != null, "result must not be null");
-
-            if (forceRecompile && !result.PreserveDetailsWhenForceRecompile)
-            {
-                return CreateForceCompileResult(result);
-            }
-
-            if (result.IsIndeterminate)
-            {
-                return new CompileResponse(
-                    success: result.Success,
-                    errorCount: result.ErrorCount,
-                    warningCount: result.WarningCount,
-                    errors: null,
-                    warnings: null,
-                    message: result.Message ?? "Compilation status is unknown. Use get-logs to inspect the compiler output.");
-            }
-
-            return new CompileResponse(
-                success: result.Success,
-                errorCount: result.Errors?.Length ?? 0,
-                warningCount: result.Warnings?.Length ?? 0,
-                errors: ToIssues(result.Errors),
-                warnings: ToIssues(result.Warnings),
-                message: AddMissingTestFrameworkReferenceHint(result.Message, result.Errors));
-        }
-
-        internal static void StoreCompileResult(
-            ICompileResultSessionRepository compileResultSessionRepository,
-            IPendingCompileSessionRepository pendingCompileSessionRepository,
-            string requestId,
-            bool forceRecompile,
-            CompileResponse result,
-            string correlationId)
-        {
-            Debug.Assert(compileResultSessionRepository != null, "compileResultSessionRepository must not be null");
-            Debug.Assert(pendingCompileSessionRepository != null, "pendingCompileSessionRepository must not be null");
-            Debug.Assert(!string.IsNullOrWhiteSpace(requestId), "requestId must not be null or whitespace");
-            Debug.Assert(result != null, "result must not be null");
-
-            ICompileResultSessionRepository compileResultRepository = compileResultSessionRepository ??
-                throw new ArgumentNullException(nameof(compileResultSessionRepository));
-            IPendingCompileSessionRepository pendingCompileRepository = pendingCompileSessionRepository ??
-                throw new ArgumentNullException(nameof(pendingCompileSessionRepository));
-            result.ProjectRoot = UnityCliLoopPathResolver.GetProjectRoot();
-            string resultJson = JsonConvert.SerializeObject(
-                result,
-                Formatting.None,
-                JsonRpcResponseSerializer.Settings);
-            UnityCliLoopStoredCompileResult previousResult =
-                compileResultRepository.GetCompileResult(requestId);
-            UnityCliLoopPendingCompileRequest pendingRequest =
-                pendingCompileRepository.GetPendingCompileRequestForRequestId(requestId);
-            compileResultRepository.StoreCompileResult(
-                requestId,
-                forceRecompile,
-                resultJson,
-                DateTime.UtcNow);
-            bool pendingRequestCleared =
-                pendingCompileRepository.ClearPendingCompileRequestIfMatches(requestId);
-            VibeLogger.LogInfo(
-                "compile_result_session_state_store_complete",
-                "Stored compile result in SessionState for CLI status polling.",
-                new
-                {
-                    request_id = requestId,
-                    force_recompile = forceRecompile,
-                    success = result.Success,
-                    error_count = result.ErrorCount,
-                    warning_count = result.WarningCount,
-                    result_bytes = System.Text.Encoding.UTF8.GetByteCount(resultJson),
-                    store_sequence = previousResult.HasResult ? 2 : 1,
-                    pending_request_before = pendingRequest.HasRequest,
-                    pending_request_cleared = pendingRequestCleared,
-                    duplicate_result_for_request = previousResult.HasResult
-                },
-                correlationId);
-        }
-
-        private static CompileResponse CreateForceCompileResult(CompileResult result)
-        {
-            ForceCompileUnknownResult unknownResult = ForceCompileUnknownResult.Create(result.Success);
-            return new CompileResponse(
-                success: unknownResult.Success,
-                errorCount: unknownResult.ErrorCount,
-                warningCount: unknownResult.WarningCount,
-                errors: null,
-                warnings: null,
-                message: unknownResult.Message);
-        }
-
-        private static CompileIssue[] ToIssues(UnityEditor.Compilation.CompilerMessage[] messages)
-        {
-            if (messages == null)
-            {
-                return null;
-            }
-
-            return messages.Select(message => new CompileIssue(message.message, message.file, message.line)).ToArray();
-        }
-
-        private static string AddMissingTestFrameworkReferenceHint(
-            string message,
-            UnityEditor.Compilation.CompilerMessage[] errors)
-        {
-            if (!ContainsMissingTestFrameworkReference(errors))
-            {
-                return message;
-            }
-
-            if (string.IsNullOrWhiteSpace(message))
-            {
-                return MissingTestFrameworkReferenceHint;
-            }
-
-            if (message.Contains(MissingTestFrameworkReferenceHint, StringComparison.Ordinal))
-            {
-                return message;
-            }
-
-            return $"{message} {MissingTestFrameworkReferenceHint}";
-        }
-
-        private static bool ContainsMissingTestFrameworkReference(
-            UnityEditor.Compilation.CompilerMessage[] errors)
-        {
-            if (errors == null)
-            {
-                return false;
-            }
-
-            foreach (UnityEditor.Compilation.CompilerMessage error in errors)
-            {
-                string errorMessage = error.message ?? string.Empty;
-                if (errorMessage.Contains("UnityTestAttribute", StringComparison.Ordinal)
-                    || errorMessage.Contains("UnityEngine.TestTools", StringComparison.Ordinal)
-                    || errorMessage.Contains("NUnit.Framework", StringComparison.Ordinal))
-                {
-                    return true;
-                }
-            }
-
-            return false;
-        }
-    }
-
-    /// <summary>
-    /// Describes whether a compile controller should record its result for delayed CLI polling.
-    /// </summary>
-    internal readonly struct CompileResultRecordingContext
-    {
-        private CompileResultRecordingContext(bool enabled, string requestId, bool forceRecompile)
-        {
-            Enabled = enabled;
-            RequestId = requestId;
-            ForceRecompile = forceRecompile;
-        }
-
-        internal bool Enabled { get; }
-        internal string RequestId { get; }
-        internal bool ForceRecompile { get; }
-
-        internal static CompileResultRecordingContext Disabled()
-        {
-            return new CompileResultRecordingContext(false, "", false);
-        }
-
-        internal static CompileResultRecordingContext Create(CompileSchema request)
-        {
-            Debug.Assert(request != null, "request must not be null");
-
-            if (!request.WaitForDomainReload || string.IsNullOrWhiteSpace(request.RequestId))
-            {
-                return Disabled();
-            }
-
-            return new CompileResultRecordingContext(
-                true,
-                request.RequestId,
-                request.ForceRecompile);
-        }
-    }
-}
