+++ Packages/src/Editor/FirstPartyTools/Compile/CompileSessionResultStore.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3636ba4c6b144241b76fa04c25dd2fa6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
